Move packet logic to packet.rs
